requirement.txt
numpy
==1.16.2
torch>=1.0

torchvision==0.2.2

pycocotools

pillow